backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
clerk-sdk-python
openai
pydantic
langgraph
langchain
langchain-openai
langchain-core
langchain-anthropic
